+++ pycaret/classification.py
@@ -2,7 +2,7 @@
 # Author: Moez Ali <moez.ali@queensu.ca>
 # License: MIT
 # Release: PyCaret 2.0x
-# Last modified : 29/07/2020
+# Last modified : 30/07/2020
 
 def setup(data,  
           target,   
@@ -9090,6 +9090,13 @@ def interpret_model(estimator,
 
     logger.info("Checking exceptions")
 
+    #checking if shap available
+    try:
+        import shap
+    except:
+        logger.error("shap library not found. pip install shap to use interpret_model function.")
+        sys.exit("shap library not found. pip install shap to use interpret_model function.")   
+
     #allowed models
     allowed_models = ['RandomForestClassifier',
                       'DecisionTreeClassifier',
@@ -11309,6 +11316,12 @@ def deploy_model(model,
     logger.info("""deploy_model(model={}, model_name={}, authentication={}, platform={})""".\
         format(str(model), str(model_name), str(authentication), str(platform)))
 
+    #checking if awscli available
+    try:
+        import awscli
+    except:
+        logger.error("awscli library not found. pip install awscli to use deploy_model function.")
+        sys.exit("awscli library not found. pip install awscli to use deploy_model function.")  
 
     #ignore warnings
     import warnings