+++ pycaret/regression.py
@@ -2,7 +2,7 @@
 # Author: Moez Ali <moez.ali@queensu.ca>
 # License: MIT
 # Release: PyCaret 2.0x
-# Last modified : 29/07/2020
+# Last modified : 30/07/2020
 
 def setup(data, 
           target, 
@@ -8588,6 +8588,13 @@ def interpret_model(estimator,
 
     logger.info("Checking exceptions")
 
+    #checking if shap available
+    try:
+        import shap
+    except:
+        logger.error("shap library not found. pip install shap to use interpret_model function.")
+        sys.exit("shap library not found. pip install shap to use interpret_model function.")  
+        
     #allowed models
     allowed_models = ['RandomForestRegressor',
                       'DecisionTreeRegressor',
@@ -9868,6 +9875,13 @@ def deploy_model(model,
     logger.info("Initializing deploy_model()")
     logger.info("""deploy_model(model={}, model_name={}, authentication={}, platform={})""".\
         format(str(model), str(model_name), str(authentication), str(platform)))
+
+    #checking if awscli available
+    try:
+        import awscli
+    except:
+        logger.error("awscli library not found. pip install awscli to use deploy_model function.")
+        sys.exit("awscli library not found. pip install awscli to use deploy_model function.")  
 
     #ignore warnings
     import warnings